nur: add eval command for local testing
