+++ ci/nur/update.py
@@ -1,62 +1,12 @@
 import logging
-import os
-import subprocess
-import tempfile
 from argparse import Namespace
-from pathlib import Path
 
-from .error import EvalError
+from .eval import EvalError, eval_repo
 from .manifest import Repo, load_manifest, update_lock_file
-from .path import EVALREPO_PATH, LOCK_PATH, MANIFEST_PATH, nixpkgs_path
+from .path import LOCK_PATH, MANIFEST_PATH
 from .prefetch import prefetch
 
 logger = logging.getLogger(__name__)
-
-
-def eval_repo(repo: Repo, repo_path: Path) -> None:
-    with tempfile.TemporaryDirectory() as d:
-        eval_path = Path(d).joinpath("default.nix")
-        with open(eval_path, "w") as f:
-            f.write(
-                f"""
-                    with import <nixpkgs> {{}};
-import {EVALREPO_PATH} {{
-  name = "{repo.name}";
-  url = "{repo.url}";
-  src = {repo_path.joinpath(repo.file)};
-  inherit pkgs lib;
-}}
-"""
-            )
-
-        # fmt: off
-        cmd = [
-            "nix-env",
-            "-f", str(eval_path),
-            "-qa", "*",
-            "--meta",
-            "--xml",
-            "--allowed-uris", "https://static.rust-lang.org",
-            "--option", "restrict-eval", "true",
-            "--option", "allow-import-from-derivation", "true",
-            "--drv-path",
-            "--show-trace",
-            "-I", f"nixpkgs={nixpkgs_path()}",
-            "-I", str(repo_path),
-            "-I", str(eval_path),
-            "-I", str(EVALREPO_PATH),
-        ]
-        # fmt: on
-
-        logger.info(f"Evaluate repository {repo.name}")
-        env = dict(PATH=os.environ["PATH"], NIXPKGS_ALLOW_UNSUPPORTED_SYSTEM="1")
-        proc = subprocess.Popen(cmd, env=env, stdout=subprocess.DEVNULL)
-        try:
-            res = proc.wait(15)
-        except subprocess.TimeoutExpired:
-            raise EvalError(f"evaluation for {repo.name} timed out of after 15 seconds")
-        if res != 0:
-            raise EvalError(f"{repo.name} does not evaluate:\n$ {' '.join(cmd)}")
 
 
 def update(repo: Repo) -> Repo:
